menghapus edit laporan di superadmin

+++ app/Views/admin/reportlaporan.php
@@ -62,13 +62,23 @@
                         <td><?= $r['lokasi_kejadian']; ?></td>
                         <td><?= $r['tindak_lanjut']; ?></td>
                         <!-- <td><a href="/report/download/<?= $r['id_pelapor']; ?>" class="btn btn-success btn-sm">Download</a></td> -->
-                        <td>
-                          <a href="/report/detailReport/<?= $r['id_pelapor']; ?>" class="btn btn-info btn-sm">Detail</a>
-                          <a href="/report/editReport/<?= $r['id_pelapor']; ?>" class="btn btn-warning btn-sm">Edit</a>
-                          <a href="/report/delete/<?= $r['id_pelapor']; ?>" class="btn btn-danger btn-sm" onclick="return confirm('Yakin data ini akan dihapus');">Hapus</a>
-                        </td>
+                        <?php
+                        if (session()->get('level') == "Super Admin") {
+                        ?>
+                          <td>
+                            <a href="/report/detailReport/<?= $r['id_pelapor']; ?>" class="btn btn-info btn-sm">Detail</a>
+                            <a href="/report/delete/<?= $r['id_pelapor']; ?>" class="btn btn-danger btn-sm" onclick="return confirm('Yakin data ini akan dihapus');">Hapus</a>
+                          </td>
+                        <?php
+                        } else if (session()->get('level') == "Admin") {
+                        ?>
+                          <td>
+                            <a href="/report/detailReport/<?= $r['id_pelapor']; ?>" class="btn btn-info btn-sm">Detail</a>
+                            <a href="/report/editReport/<?= $r['id_pelapor']; ?>" class="btn btn-warning btn-sm">Edit</a>
+                            <a href="/report/delete/<?= $r['id_pelapor']; ?>" class="btn btn-danger btn-sm" onclick="return confirm('Yakin data ini akan dihapus');">Hapus</a>
+                          </td>
+                        <?php } ?>
                       </tr>
-
                     <?php
                       $no++;
                     endforeach; ?>
